FIXED THE GC CRASHES! It's working here.
Need to cleanup code, as I added printf calls at each stage of GC to see where the root of problem was. The problem was caused by the fact that there is memory allocated elsewhere in Julia that is wrapped around a jl_ptr_to_array_1d (or multidimensional aswell, need more checks) with the own_buffer flag set to 1. This flag would then allow the GC to call free() on that memory. I have a horrible workaround here that would #undef the free that relies on RTFree, and #define it back again after a standard free() call to that external malloced memory.